# examples/demo-backend/openapi/complex.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/conversations": {
      "get": {
        "operationId": "Chat_list",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "List conversations available to the current user",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/conversations/{id}": {
      "get": {
        "operationId": "Chat_get",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "Get one conversation",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/conversations/{id}/messages": {
      "get": {
        "operationId": "Chat_messages",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "example": "notification-020",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "Load message history using cursor pagination",
        "tags": [
          "Chat"
        ]
      },
      "post": {
        "operationId": "Chat_send",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "csrf": []
          }
        ],
        "summary": "Send an idempotent message through REST and broadcast it over Socket.IO",
        "tags": [
          "Chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "Demo Complex API",
    "description": "Cookie-session, multitenant and realtime API for large frontend applications.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "demo_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required for authenticated mutations."
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "email"
          },
          "message": {
            "type": "string",
            "example": "must be an email"
          },
          "code": {
            "type": "string",
            "example": "isEmail"
          }
        },
        "required": [
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "PRODUCT_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Product not found"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-30T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products/product-404"
          },
          "requestId": {
            "type": "string",
            "example": "req-5c9f7a3d"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "timestamp",
          "path",
          "requestId"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CursorMetaDto": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number",
            "example": 20,
            "minimum": 1
          },
          "nextCursor": {
            "type": "object",
            "nullable": true,
            "example": "product-020"
          },
          "hasMore": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "limit",
          "hasMore"
        ]
      },
...
      "ConversationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "conversation-support"
          },
          "organizationId": {
            "type": "string",
            "example": "org-acme"
          },
          "title": {
            "type": "string",
            "example": "Order support"
          },
          "participantIds": {
            "example": [
              "complex-user-admin",
              "complex-user-support"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastMessagePreview": {
            "type": "object",
            "nullable": true,
            "example": "Can you check order complex-order-001?"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organizationId",
          "title",
          "participantIds",
          "lastMessagePreview",
          "updatedAt"
        ]
      },
      "ConversationsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "ConversationResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConversationDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "ChatMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "message-001"
          },
          "conversationId": {
            "type": "string",
            "example": "conversation-support"
          },
          "senderId": {
            "type": "string",
            "example": "complex-user-support"
          },
          "text": {
            "type": "string",
            "example": "The order has already been packed."
          },
          "clientMessageId": {
            "type": "string",
            "example": "client-message-4c08",
            "description": "Frontend-generated key used to deduplicate retries."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "conversationId",
          "senderId",
          "text",
          "clientMessageId",
          "createdAt"
        ]
      },
      "MessagesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/CursorMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "The order has already been packed."
          },
          "clientMessageId": {
            "type": "string",
            "example": "client-message-4c08"
          }
        },
        "required": [
          "text",
          "clientMessageId"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ChatMessageDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}